Refactor discord bot
Made discord bot more resilient with the power of executors and asyncio.
Now it can process multiple concurrent  requests.
