+++ src/bots/discord/__init__.py
@@ -1,0 +1,1 @@
+from .discord_bot import DiscordBot
